Extract toggles and toggle-sets to aiotoggles
