+++ kopf/reactor/observation.py
@@ -25,9 +25,10 @@ import functools
 import logging
 from typing import Collection, List, Optional
 
+from kopf.aiokits import aiotoggles
 from kopf.clients import errors, fetching, scanning
 from kopf.reactor import queueing, registries
-from kopf.structs import bodies, configuration, handlers, primitives, references
+from kopf.structs import bodies, configuration, handlers, references
 
 logger = logging.getLogger(__name__)
 
@@ -139,8 +140,8 @@ async def process_discovered_namespace_event(
         insights: references.Insights,
         # Must be accepted whether used or not -- as passed by watcher()/worker().
         stream_pressure: Optional[asyncio.Event] = None,  # None for tests
-        resource_indexed: Optional[primitives.Toggle] = None,  # None for tests & observation
-        operator_indexed: Optional[primitives.ToggleSet] = None,  # None for tests & observation
+        resource_indexed: Optional[aiotoggles.Toggle] = None,  # None for tests & observation
+        operator_indexed: Optional[aiotoggles.ToggleSet] = None,  # None for tests & observation
 ) -> None:
     if raw_event['type'] is None:
         return
@@ -157,8 +158,8 @@ async def process_discovered_resource_event(
         insights: references.Insights,
         # Must be accepted whether used or not -- as passed by watcher()/worker().
         stream_pressure: Optional[asyncio.Event] = None,  # None for tests
-        resource_indexed: Optional[primitives.Toggle] = None,  # None for tests & observation
-        operator_indexed: Optional[primitives.ToggleSet] = None,  # None for tests & observation
+        resource_indexed: Optional[aiotoggles.Toggle] = None,  # None for tests & observation
+        operator_indexed: Optional[aiotoggles.ToggleSet] = None,  # None for tests & observation
 ) -> None:
     # Ignore the initial listing, as all custom resources were already noticed by API listing.
     # This prevents numerous unneccessary API requests at the the start of the operator.
