+++ tests/k8s/test_watching_with_freezes.py
@@ -4,8 +4,8 @@ import logging
 import async_timeout
 import pytest
 
+from kopf.aiokits.aiotoggles import ToggleSet
 from kopf.clients.watching import streaming_block
-from kopf.structs.primitives import ToggleSet
 
 
 async def test_pausing_is_ignored_if_turned_off(
